+++ common/eth2/src/lighthouse.rs
@@ -56,18 +56,18 @@ pub struct Peer<T: EthSpec> {
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct GlobalValidatorInclusionData {
     /// The total effective balance of all active validators during the _current_ epoch.
-    pub current_epoch_active_gwei: u64,
+    pub current_epoch_active_gwei: u128,
     /// The total effective balance of all active validators during the _previous_ epoch.
-    pub previous_epoch_active_gwei: u64,
+    pub previous_epoch_active_gwei: u128,
     /// The total effective balance of all validators who attested during the _current_ epoch and
     /// agreed with the state about the beacon block at the first slot of the _current_ epoch.
-    pub current_epoch_target_attesting_gwei: u64,
+    pub current_epoch_target_attesting_gwei: u128,
     /// The total effective balance of all validators who attested during the _previous_ epoch and
     /// agreed with the state about the beacon block at the first slot of the _previous_ epoch.
-    pub previous_epoch_target_attesting_gwei: u64,
+    pub previous_epoch_target_attesting_gwei: u128,
     /// The total effective balance of all validators who attested during the _previous_ epoch and
     /// agreed with the state about the beacon block at the time of attestation.
-    pub previous_epoch_head_attesting_gwei: u64,
+    pub previous_epoch_head_attesting_gwei: u128,
 }
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
