+++ consensus/fork_choice/src/fork_choice.rs
@@ -1273,7 +1273,7 @@ where
     }
 
     /// Returns the weight for the given block root.
-    pub fn get_block_weight(&self, block_root: &Hash256) -> Option<u64> {
+    pub fn get_block_weight(&self, block_root: &Hash256) -> Option<u128> {
         self.proto_array.get_weight(block_root)
     }
