+++ consensus/state_processing/src/per_block_processing/altair/sync_committee.rs
@@ -71,10 +71,10 @@ pub fn compute_sync_aggregate_rewards<T: EthSpec>(
 ) -> Result<(u64, u64), BlockProcessingError> {
     let total_active_balance = state.get_total_active_balance()?;
     let total_active_increments =
-        total_active_balance.safe_div(spec.effective_balance_increment)?;
+        total_active_balance.safe_div(spec.effective_balance_increment as u128)?;
     let total_base_rewards = BaseRewardPerIncrement::new(total_active_balance, spec)?
         .as_u64()
-        .safe_mul(total_active_increments)?;
+        .safe_mul(total_active_increments as u64)?;
     let max_participant_rewards = total_base_rewards
         .safe_mul(SYNC_REWARD_WEIGHT)?
         .safe_div(WEIGHT_DENOMINATOR)?
