+++ consensus/state_processing/src/per_epoch_processing.rs
@@ -47,19 +47,19 @@ pub fn process_epoch<T: EthSpec>(
 /// Used to track the changes to a validator's balance.
 #[derive(Default, Clone)]
 pub struct Delta {
-    pub rewards: u64,
-    pub penalties: u64,
+    pub rewards: u128,
+    pub penalties: u128,
 }
 
 impl Delta {
     /// Reward the validator with the `reward`.
-    pub fn reward(&mut self, reward: u64) -> Result<(), Error> {
+    pub fn reward(&mut self, reward: u128) -> Result<(), Error> {
         self.rewards = self.rewards.safe_add(reward)?;
         Ok(())
     }
 
     /// Penalize the validator with the `penalty`.
-    pub fn penalize(&mut self, penalty: u64) -> Result<(), Error> {
+    pub fn penalize(&mut self, penalty: u128) -> Result<(), Error> {
         self.penalties = self.penalties.safe_add(penalty)?;
         Ok(())
     }
